Added additional pages to the site, updated index and meet the team.
Removed meet the team from the index and created a page for its self,
with temp  images for each teammate. Added a similar setup for the
sponsors and the mentor. Created pages for the project and for
documentation with "Under Construction" showing on both. Set up the nav
bar on all pages to work between all pages.