Introduced favicon to all pages, updated the documents page to include links to all pdfs of completed documents.
Created favicon icons for the site with the logo, there are unique
favicon for each type of device. Used javascrip to detect the device,
created using a tool. Added the code for favicons to each page. Created
link buttons on the documents page, to the completed documetns. Also
created inactive links buttons to expected future documents.